+++ src/Agents.Net/Agent.cs
@@ -11,13 +11,13 @@ using System;
 using System.Collections.Generic;
 using System.Runtime.ExceptionServices;
 using System.Text;
-using NLog;
+using Serilog;
+using Serilog.Events;
 
 namespace Agents.Net
 {
     public abstract class Agent
     {
-        private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
         private readonly IMessageBoard messageBoard;
         private readonly string agentName;
 
@@ -27,13 +27,20 @@ namespace Agents.Net
             agentName = string.IsNullOrEmpty(name) ? GetType().Name : name;
         }
 
+        protected Guid Id { get; } = Guid.NewGuid();
+
         public void Execute(Message messageData)
         {
             if (messageData == null)
             {
                 throw new ArgumentNullException(nameof(messageData));
             }
-            Logger.Trace(new AgentLog(messageData, "Executing", agentName));
+
+            if (Log.IsEnabled(LogEventLevel.Verbose))
+            {
+                Log.Verbose("{@log}",
+                            new AgentLog(agentName, "Executing", Id, messageData.ToMessageLog()));
+            }
 
             try
             {
@@ -48,7 +55,11 @@ namespace Agents.Net
 
         protected void OnMessage(Message message)
         {
-            Logger.Trace(new AgentLog(message, "Publishing", agentName));
+            if (Log.IsEnabled(LogEventLevel.Verbose))
+            {
+                Log.Verbose("{@log}",
+                            new AgentLog(agentName, "Publishing", Id, message?.ToMessageLog()));
+            }
             messageBoard.Publish(message);
         }
 
@@ -70,31 +81,5 @@ namespace Agents.Net
         }
 
         protected abstract void ExecuteCore(Message messageData);
-
-        protected class AgentLog
-        {
-            private readonly Message message;
-            private readonly string type;
-            private readonly string agentName;
-
-            public AgentLog(Message message, string type, string agentName)
-            {
-                this.message = message;
-                this.type = type;
-                this.agentName = agentName;
-            }
-
-            public override string ToString()
-            {
-                StringBuilder messageBuilder = new StringBuilder("{\"Message\": ");
-                messageBuilder.Append(message.ToStringBuilder());
-                messageBuilder.Append(", \"Agent\": \"");
-                messageBuilder.Append(agentName);
-                messageBuilder.Append("\", \"Type\": \"");
-                messageBuilder.Append(type);
-                messageBuilder.Append("\"}");
-                return messageBuilder.ToString();
-            }
-        }
     }
 }