+++ src/Agents.Net/MessageBoard.cs
@@ -12,7 +12,6 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
-using NLog;
 
 namespace Agents.Net
 {