+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 - **Breaking Change:** Switch from magic string based agent definition to type based agent definition
 - **Breaking Change:** Change execution method of aggregator to `IReadonlyCollection` so that it can be passed directly to TerminateDomainsOf
+- **Breaking Change:** Switched logging framework from NLog to Serilog
 - `MessageCollector` messages can know be consumed directly, so that they are removed from the collector
 - `Message` predecessor property visibility changed to public